renamed ppu_ operations and made explicit dword write

+++ core/timer.cpp
@@ -15,15 +15,15 @@ s8 Timer::tima_inc() {
     */
     if (mem.read(0xFF05) == 0xFF) {
         tima_flag = true;
-        mem.write(0xFF05, (u8)0x00);
+        mem.write(0xFF05, 0x00);
         tima_val = mem.read(0xFF06);
         return -1;
     } else if ((mem.read(0xFF05) == 0) && tima_flag) {
         tima_flag = false;
-        mem.write(0xFF05, (u8)(tima_val));
-        mem.write(0xFF0F, (u8)(mem.read(0xFF0F) | 0b100));
+        mem.write(0xFF05, (tima_val));
+        mem.write(0xFF0F, (mem.read(0xFF0F) | 0b100));
     } else {
-        mem.write(0xFF05, (u8)(mem.read(0xFF05) + 1));
+        mem.write(0xFF05, (mem.read(0xFF05) + 1));
     }
     return 0;
 }
